+++ main.go
@@ -16,10 +16,8 @@ package main
 
 import (
 	"gitlab.com/cogment/cogment/cmd"
-	"log"
 )
 
 func main() {
-	log.SetFlags(0)
 	cmd.Execute()
 }