+++ CHANGELOG.md
@@ -7,6 +7,47 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## v1.0.0-alpha4 - 2021-02-19
+
+### Features
+
+#### implement web-client
+
+This implements both the creation and generation of a web-client through
+`cogment init` and `cogment generate`, respectively.
+
+**Changes**:
+
+- Refactor code to use template helpers for transforming `cogment.yaml`
+  `import.protos[]` into language specific import strings.
+
+**Breaking Changes**
+
+- A node.js distribution is now required to be available on `$PATH` for
+  certain features. These are disabled by default and must be enabled by
+  stdin or flags.
+- `cogment generate` now accepts `--python-out` instead of
+  `--python_out`. `--python-out` can be repeated multiple times to
+  target multiple output directories, eg: `cogment generate --python-out
+  environment --python-out client --python-out actor`.
+- `cogment generate` now templates `CogSettings.ts`, then uses the
+  typescript compiler to generate `CogSettings.js`, `CogSettings.d.ts`,
+  `CogSettings.d.ts.map`.
+- `cogment init` now asks if a web-client should be generated.
+- `cogment generate` now accepts the `--web-client` flag, which enables
+  generation of protobuf definitions and `CogSettings.ts` from a
+  `cogment.yaml`
+- `cogment generate` now accepts a `--typescript` flag that depends on
+  the `--web-client` flag, will enable typing definition generation for
+  user protobufs.
+- `cogment init` templates uses dependencies between services
+  (`depends_on` clause) for bringing up the stack vs. having service
+  names repeated in `cogment run` commands. `docker-compose up
+  web-client` will bring up the entire stack, `docker-compose up
+  dashboard` will bring up all the necessary containers.
+
+
+
 ## v1.0.0-alpha3 - 2021-02-19
 
 ### Changed